+++ src/pages/CompanyIntroduction/pages/Location.tsx
@@ -1,0 +1,25 @@
+import PageLayout from '@/components/organisms/PageLayout/PageLayout';
+import contactUs from '/ContactUs_Image.svg';
+import LocationContent from '../components/LocationContent';
+
+function Location() {
+  return (
+    <>
+      <PageLayout
+        isMainMenu={false}
+        EngTitle="CONTACT US"
+        KorTitle="찾아오시는 길"
+        alt="찾아오시는 길 페이지"
+        mainImage={contactUs}
+        firstTitle="CONTACT"
+        firstSubTitle="US"
+        secondTitle="찾아오시는 길"
+        thirdTitle="주식회사 나람에 찾아오시는 길입니다 "
+      >
+        <LocationContent />
+      </PageLayout>
+    </>
+  );
+}
+
+export default Location;
