--- src/pages/Introduction/components/LocationContent.tsx
@@ -1,21 +1,0 @@
-function LocationContent() {
-  return (
-    <>
-      <div>
-        찾아오시는 길
-        페이지ddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd시는 길
-        페이지dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd
-      </div>
-    </>
-  );
-}
-
-export default LocationContent;
